move backup integration tests to tests
