+++ components/backup/Cargo.toml
@@ -4,10 +4,6 @@ version = "0.0.1"
 edition = "2018"
 publish = false
 
-[[test]]
-name = "integrations"
-path = "tests/integrations/mod.rs"
-
 [features]
 default = ["protobuf-codec", "test-engines-rocksdb"]
 protobuf-codec = [
@@ -78,9 +74,5 @@ txn_types = { path = "../txn_types" }
 yatp = { git = "https://github.com/tikv/yatp.git", branch = "master" }
 
 [dev-dependencies]
-futures-executor = "0.3"
 rand = "0.7"
 tempfile = "3.0"
-test_util = { path = "../test_util" }
-test_raftstore = { path = "../test_raftstore" }
-uuid = { version = "0.8.1", features = ["serde", "v4"] }
