+++ tests/integrations/mod.rs
@@ -12,6 +12,7 @@ extern crate encryption;
 extern crate tikv_util;
 extern crate pd_client;
 
+mod backup;
 mod config;
 mod coprocessor;
 mod import;
